move poke agents/skills API contracts out of pages/api
