+++ front/pages/api/poke/workspaces/[wId]/skill_suggestions/[suggestionId]/details.ts
@@ -1,6 +1,7 @@
 /** @ignoreswagger */
 // @migration-status: MIGRATED_TO_HONO
 import { withSessionAuthenticationForPoke } from "@app/lib/api/auth_wrappers";
+import type { PokeGetSkillSuggestionDetails } from "@app/lib/api/poke/skills";
 import { Authenticator } from "@app/lib/auth";
 import type { SessionWithUser } from "@app/lib/iam/provider";
 import { SkillResource } from "@app/lib/resources/skill/skill_resource";
@@ -8,14 +9,7 @@ import { SkillSuggestionResource } from "@app/lib/resources/skill_suggestion_res
 import { apiError } from "@app/logger/withlogging";
 import type { WithAPIErrorResponse } from "@app/types/error";
 import { isString } from "@app/types/shared/utils/general";
-import type { SkillSuggestionType } from "@app/types/suggestions/skill_suggestion";
 import type { NextApiRequest, NextApiResponse } from "next";
-
-export type PokeGetSkillSuggestionDetails = {
-  suggestion: SkillSuggestionType;
-  skillInstructionsHtml: string | null;
-  skillAgentFacingDescription: string | null;
-};
 
 async function handler(
   req: NextApiRequest,
