+++ front/pages/api/poke/workspaces/[wId]/skills/[sId]/suggestions.ts
@@ -1,19 +1,15 @@
 /** @ignoreswagger */
 // @migration-status: MIGRATED_TO_HONO
 import { withSessionAuthenticationForPoke } from "@app/lib/api/auth_wrappers";
+import type { PokeListSkillSuggestions } from "@app/lib/api/poke/skills";
 import { Authenticator } from "@app/lib/auth";
 import type { SessionWithUser } from "@app/lib/iam/provider";
 import { SkillSuggestionResource } from "@app/lib/resources/skill_suggestion_resource";
 import { apiError } from "@app/logger/withlogging";
 import type { WithAPIErrorResponse } from "@app/types/error";
 import { isString } from "@app/types/shared/utils/general";
-import type { SkillSuggestionType } from "@app/types/suggestions/skill_suggestion";
 import { SKILL_SUGGESTION_SOURCES } from "@app/types/suggestions/skill_suggestion";
 import type { NextApiRequest, NextApiResponse } from "next";
-
-export type PokeListSkillSuggestions = {
-  suggestions: SkillSuggestionType[];
-};
 
 async function handler(
   req: NextApiRequest,
