+++ front/poke/swr/skill_details.ts
@@ -1,6 +1,8 @@
+import type {
+  PokeGetSkillDetails,
+  PokeGetSkillVersions,
+} from "@app/lib/api/poke/skills";
 import { emptyArray, useFetcher, useSWRWithDefaults } from "@app/lib/swr/swr";
-import type { PokeGetSkillDetails } from "@app/pages/api/poke/workspaces/[wId]/skills/[sId]/details";
-import type { PokeGetSkillVersions } from "@app/pages/api/poke/workspaces/[wId]/skills/[sId]/versions";
 import type { LightWorkspaceType } from "@app/types/user";
 import type { Fetcher } from "swr";
