+++ front/poke/swr/skill_suggestion_details.ts
@@ -1,5 +1,5 @@
+import type { PokeGetSkillSuggestionDetails } from "@app/lib/api/poke/skills";
 import { useFetcher, useSWRWithDefaults } from "@app/lib/swr/swr";
-import type { PokeGetSkillSuggestionDetails } from "@app/pages/api/poke/workspaces/[wId]/skill_suggestions/[suggestionId]/details";
 import type { LightWorkspaceType } from "@app/types/user";
 import type { Fetcher } from "swr";
